pyproject.toml
[project]
name = "inventario-ghl"
version = "0.1.0"
description = "Aplicación de escritorio para generar reportes de inventario de HighLevel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.7.0",
    "xlsxwriter>=3.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "cx-freeze>=6.15.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py310']